Fix migration when dest is on different partition
os.Rename will fail if the destination file/folder is on a different
partition from the source. Unfortunately, in this case the only option
is to copy the file or each of the folder's files then delete the
original. Update migrate.go to handle this case correctly.

For the folder-copying case, I have used this library here: https://github.com/otiai10/copy